fix line_append and text_append
crashed on cygwin because of a double-free. still leaking a bit of
memory which is actually the entire contents of the file you edit. not a
good thing at all, something to do with realloc? the pointers are
completely lost by the time valgrind notices them which is annoying.
whatever, it's christmas tomorrow and I don't feel morally obligated to
continue working on this project. this is a long ass-commit message. I'd
like to thank my mom, my dad, stack overflow, and the internet for
making this project possible; in the HALL. HAL. space oddessey (that's
probably misspelled) 2001. I crashed valgrind and it told me to fix my
program before submitting a bug report. I thought that was pretty funny.
we can keep going, but it seems like I'm already resorting to
motivational bullshitting so I don't think it's likely as my brain is
much right now but typing on this new mechanical keyboard is so fun and
I just want to mush my face all over it. need some stabs for the space
bar though. annoying sometimes. I should really push this commit. bye